- Makefile: Revamped VERBOSE support, build mechanics changed.   - VERBOSE has now greater effect, many previously hidden commands     are now shown on the terminal as they are executed. (Use $(Q)     instead of @ where the command should be shown if VERBOSE.)   - Compilation commands redundancy was eliminated.   - Compilation commands calling generalized so that the VERBOSE     support does not further entangle the code. The commands are now     all at a single place, before the targets in a "Build commands"     section.
Note that this changeset has a character of proposal. If you do not
like the direction it takes at all, feel free to revert it (probably
together with the previous VERBOSE patch). However, I believe it
improves the code by bringing the compilation commands to a single
place and it greatly improves the VERBOSE support, too (and makes it
much less intrusive).



git-svn-id: http://svn.openttd.org/trunk@1917 6aa0318a-3be1-0310-93fa-89fd2396df07